feat(zendesk): separates zendesk/user route logic into separate func
